tests: Resolve C4267 warning on MSVC

+++ src/tests/core/hle/kernel/hle_ipc.cpp
@@ -142,8 +142,8 @@ TEST_CASE("HLERequestContext::PopulateFromIncomingCommandBuffer", "[core][kernel
         std::fill(buffer.GetPtr(), buffer.GetPtr() + buffer.GetSize(), 0xAB);
 
         VAddr target_address = 0x10000000;
-        auto result = process->vm_manager.MapBackingMemory(target_address, buffer, buffer.GetSize(),
-                                                           MemoryState::Private);
+        auto result = process->vm_manager.MapBackingMemory(
+            target_address, buffer, static_cast<u32>(buffer.GetSize()), MemoryState::Private);
         REQUIRE(result.Code() == RESULT_SUCCESS);
 
         const u32_le input[]{
@@ -156,7 +156,8 @@ TEST_CASE("HLERequestContext::PopulateFromIncomingCommandBuffer", "[core][kernel
 
         CHECK(context.GetStaticBuffer(0) == mem->Vector());
 
-        REQUIRE(process->vm_manager.UnmapRange(target_address, buffer.GetSize()) == RESULT_SUCCESS);
+        REQUIRE(process->vm_manager.UnmapRange(
+                    target_address, static_cast<u32>(buffer.GetSize())) == RESULT_SUCCESS);
     }
 
     SECTION("translates MappedBuffer descriptors") {
@@ -165,8 +166,8 @@ TEST_CASE("HLERequestContext::PopulateFromIncomingCommandBuffer", "[core][kernel
         std::fill(buffer.GetPtr(), buffer.GetPtr() + buffer.GetSize(), 0xCD);
 
         VAddr target_address = 0x10000000;
-        auto result = process->vm_manager.MapBackingMemory(target_address, buffer, buffer.GetSize(),
-                                                           MemoryState::Private);
+        auto result = process->vm_manager.MapBackingMemory(
+            target_address, buffer, static_cast<u32>(buffer.GetSize()), MemoryState::Private);
 
         const u32_le input[]{
             IPC::MakeHeader(0, 0, 2),
@@ -181,7 +182,8 @@ TEST_CASE("HLERequestContext::PopulateFromIncomingCommandBuffer", "[core][kernel
 
         CHECK(other_buffer == mem->Vector());
 
-        REQUIRE(process->vm_manager.UnmapRange(target_address, buffer.GetSize()) == RESULT_SUCCESS);
+        REQUIRE(process->vm_manager.UnmapRange(
+                    target_address, static_cast<u32>(buffer.GetSize())) == RESULT_SUCCESS);
     }
 
     SECTION("translates mixed params") {
@@ -195,12 +197,14 @@ TEST_CASE("HLERequestContext::PopulateFromIncomingCommandBuffer", "[core][kernel
 
         VAddr target_address_static = 0x10000000;
         auto result = process->vm_manager.MapBackingMemory(
-            target_address_static, buffer_static, buffer_static.GetSize(), MemoryState::Private);
+            target_address_static, buffer_static, static_cast<u32>(buffer_static.GetSize()),
+            MemoryState::Private);
         REQUIRE(result.Code() == RESULT_SUCCESS);
 
         VAddr target_address_mapped = 0x20000000;
-        result = process->vm_manager.MapBackingMemory(
-            target_address_mapped, buffer_mapped, buffer_mapped.GetSize(), MemoryState::Private);
+        result = process->vm_manager.MapBackingMemory(target_address_mapped, buffer_mapped,
+                                                      static_cast<u32>(buffer_mapped.GetSize()),
+                                                      MemoryState::Private);
         REQUIRE(result.Code() == RESULT_SUCCESS);
 
         auto a = MakeObject(kernel);
@@ -230,9 +234,11 @@ TEST_CASE("HLERequestContext::PopulateFromIncomingCommandBuffer", "[core][kernel
         context.GetMappedBuffer(0).Read(other_buffer.data(), 0, buffer_mapped.GetSize());
         CHECK(other_buffer == mem_mapped->Vector());
 
-        REQUIRE(process->vm_manager.UnmapRange(target_address_static, buffer_static.GetSize()) ==
+        REQUIRE(process->vm_manager.UnmapRange(target_address_static,
+                                               static_cast<u32>(buffer_static.GetSize())) ==
                 RESULT_SUCCESS);
-        REQUIRE(process->vm_manager.UnmapRange(target_address_mapped, buffer_mapped.GetSize()) ==
+        REQUIRE(process->vm_manager.UnmapRange(target_address_mapped,
+                                               static_cast<u32>(buffer_mapped.GetSize())) ==
                 RESULT_SUCCESS);
     }
 }
@@ -325,7 +331,8 @@ TEST_CASE("HLERequestContext::WriteToOutgoingCommandBuffer", "[core][kernel]") {
 
         VAddr target_address = 0x10000000;
         auto result = process->vm_manager.MapBackingMemory(
-            target_address, output_buffer, output_buffer.GetSize(), MemoryState::Private);
+            target_address, output_buffer, static_cast<u32>(output_buffer.GetSize()),
+            MemoryState::Private);
         REQUIRE(result.Code() == RESULT_SUCCESS);
 
         input[0] = IPC::MakeHeader(0, 0, 2);
@@ -343,8 +350,8 @@ TEST_CASE("HLERequestContext::WriteToOutgoingCommandBuffer", "[core][kernel]") {
         context.WriteToOutgoingCommandBuffer(output_cmdbuff.data(), *process);
 
         CHECK(output_mem->Vector() == input_buffer);
-        REQUIRE(process->vm_manager.UnmapRange(target_address, output_buffer.GetSize()) ==
-                RESULT_SUCCESS);
+        REQUIRE(process->vm_manager.UnmapRange(
+                    target_address, static_cast<u32>(output_buffer.GetSize())) == RESULT_SUCCESS);
     }
 
     SECTION("translates StaticBuffer descriptors") {
@@ -356,7 +363,8 @@ TEST_CASE("HLERequestContext::WriteToOutgoingCommandBuffer", "[core][kernel]") {
 
         VAddr target_address = 0x10000000;
         auto result = process->vm_manager.MapBackingMemory(
-            target_address, output_buffer, output_buffer.GetSize(), MemoryState::Private);
+            target_address, output_buffer, static_cast<u32>(output_buffer.GetSize()),
+            MemoryState::Private);
         REQUIRE(result.Code() == RESULT_SUCCESS);
 
         const u32_le input_cmdbuff[]{
@@ -378,8 +386,8 @@ TEST_CASE("HLERequestContext::WriteToOutgoingCommandBuffer", "[core][kernel]") {
         CHECK(output[1] == IPC::MappedBufferDesc(output_buffer.GetSize(), IPC::W));
         CHECK(output[2] == target_address);
         CHECK(output_mem->Vector() == input_buffer);
-        REQUIRE(process->vm_manager.UnmapRange(target_address, output_buffer.GetSize()) ==
-                RESULT_SUCCESS);
+        REQUIRE(process->vm_manager.UnmapRange(
+                    target_address, static_cast<u32>(output_buffer.GetSize())) == RESULT_SUCCESS);
     }
 }
